+++ rankmyroast/lib/models/async_response.dart
@@ -1,0 +1,7 @@
+class AsyncResponse {
+  final bool success;
+  final bool? error;
+  final String? errorMessage;
+
+  AsyncResponse({required this.success, this.error, this.errorMessage});
+}
